requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
jinja2==3.0.1
schedule==1.1.0
gunicorn==20.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
email-validator==1.1.3
alembic==1.7.1
python-dotenv==0.19.0
pytest==6.2.5
black==21.7b0
flake8==3.9.2